refactor: move other functions to `ConstraintSettingsUtilities.ts`
